requirements.txt
fastapi==0.109.0
pydantic[email]
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.20.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
pydantic-settings==2.7.0
jwt==1.3.1
redis==4.2.0
argon2_cffi==25.1.0
celery[rabbitmq]==5.3.6
aio-pika==9.4.1
kombu==5.3.4
slowapi